Update python prebuilts
The main reason for this is b/73397411 (python spews messages about not
being able to find openssl).

Python's module build system is hideous, so after a couple of attempts I
gave up on convincing it to use a statically linked openssl (which is
how we solved the same issue for cmake prebuilts). Instead, I just chose
to hack the hashlib module to silence the errors (as we don't need
cryptographic hashing for anything we do in lldb).

Bug: 73397411
Change-Id: I28ee305d2f1519fc5e273153e097320cbe0b2f22